# openapi.json
{"openapi":"3.1.0","info":{"title":"FastAPI","version":"0.1.0"},"paths":{"/families":{"get":{"tags":["Families"],"summary":"Get All Families","operationId":"get_all_families_families_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/SafeFamily"},"type":"array","title":"Response Get All Families Families Get"}}}}}},"post":{"tags":["Families"],"summary":"Create Family","operationId":"create_family_families_post","requestBody":{"content":{"application/json":{"schema":{"type":"object","title":"Family Data"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/families/{family_id}":{"get":{"tags":["Families"],"summary":"Get Family","operationId":"get_family_families__family_id__get","parameters":[{"name":"family_id","in":"path","required":true,"schema":{"type":"string","title":"Family Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SafeFamily"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/families/search/":{"get":{"tags":["Families"],"summary":"Search Family","operationId":"search_family_families_search__get","parameters":[{"name":"group_name","in":"query","required":true,"schema":{"type":"string","title":"Group Name"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SafeFamily"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/families/{family_id}/users":{"get":{"tags":["Users"],"summary":"Get Family Users","operationId":"get_family_users_families__family_id__users_get","parameters":[{"name":"family_id","in":"path","required":true,"schema":{"type":"string","title":"Family Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/SafeUser"},"title":"Response Get Family Users Families  Family Id  Users Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/families/{family_name}/users":{"post":{"tags":["Users"],"summary":"Create Family User","operationId":"create_family_user_families__family_name__users_post","parameters":[{"name":"family_name","in":"path","required":true,"schema":{"type":"string","title":"Family Name"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","title":"User Data"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/families/{family_name}/purchases":{"post":{"tags":["Purchases"],"summary":"Create Family Purchase","operationId":"create_family_purchase_families__family_name__purchases_post","parameters":[{"name":"family_name","in":"path","required":true,"schema":{"type":"string","title":"Family Name"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","title":"Purchase Data"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/families/{family_id}/purchases":{"get":{"tags":["Purchases"],"summary":"Get Family Purchases","operationId":"get_family_purchases_families__family_id__purchases_get","parameters":[{"name":"family_id","in":"path","required":true,"schema":{"type":"string","title":"Family Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/SafePurchase"},"title":"Response Get Family Purchases Families  Family Id  Purchases Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/":{"get":{"summary":"Serve Dashboard","operationId":"serve_dashboard__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"SafeFamily":{"properties":{"id":{"type":"string","title":"Id"},"group_name":{"type":"string","title":"Group Name"},"created_at":{"type":"string","title":"Created At"},"total_balance":{"type":"number","title":"Total Balance"},"users_count":{"type":"integer","title":"Users Count"},"expenditure":{"type":"number","title":"Expenditure"},"income":{"type":"number","title":"Income"},"users":{"items":{"$ref":"#/components/schemas/SafeUser"},"type":"array","title":"Users"},"purchases":{"items":{"$ref":"#/components/schemas/SafePurchase"},"type":"array","title":"Purchases"}},"type":"object","required":["id","group_name","created_at","total_balance","users_count","expenditure","income","users","purchases"],"title":"SafeFamily"},"SafePurchase":{"properties":{"id":{"type":"string","title":"Id"},"price":{"type":"number","title":"Price"},"description":{"type":"string","title":"Description"},"user_name":{"type":"string","title":"User Name"}},"type":"object","required":["id","price","description","user_name"],"title":"SafePurchase"},"SafeUser":{"properties":{"id":{"type":"string","title":"Id"},"name":{"type":"string","title":"Name"},"surname":{"type":"string","title":"Surname"},"email":{"type":"string","format":"email","title":"Email"}},"type":"object","required":["id","name","surname","email"],"title":"SafeUser"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}